Fishy Fixes #34
intro made now hav to make game look kino

+++ Assets/Scripts/Managers/FishSellShopManager.cs
@@ -93,7 +93,7 @@ public class FishSellShopManager : MonoBehaviour
 
     public void ConfirmToSell()
     {
-        GameManager.Instance.fishCoin += FishDataManager.Instance.GetValue(selectedID) * fishStocks[selectedID].count;
+        GameManager.Instance.fishCoin += (int) FishDataManager.Instance.GetValue(selectedID) * fishStocks[selectedID].count;
 
         InventoryManager.Instance.RemoveByType(selectedID);
